backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
motor==3.1.2
pymongo==4.3.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic[email]==2.9.2
python-multipart==0.0.9
